Handle new hana_media format
Generate the new Ansible hana_media.yaml, with new format optionally
supporting token
Allow generation of the new hana_media format both from ald apiver:2 and
new config.yaml format. Only new conf.yaml allow to use tokens.
