+++ scripts/qesap/test/unit/test_qesap_configure.py
@@ -1,19 +1,6 @@
 import os
-import re
-import logging
-
-import yaml
 
 from qesap import main
-
-
-log = logging.getLogger(__name__)
-
-DATA_DIR = os.path.join(os.path.dirname(os.path.realpath(__file__)), 'assets')
-
-# @pytest.mark.datafiles(os.path.join(DATA_DIR,'repo'))
-# def test_some(datafiles):
-#    log.error(datafiles)
 
 
 def test_configure(configure_helper, config_yaml_sample):
@@ -61,395 +48,6 @@ ansible:
     args, *_ = configure_helper(provider, conf, [])
 
     assert main(args) == 1
-
-
-def test_configure_no_tfvars_template(args_helper, config_yaml_sample):
-    '''
-    if tfvars template is missing,
-    just create tfvars from the config.yaml content
-    '''
-    provider = 'pinocchio'
-    conf = config_yaml_sample(provider)
-
-    # Create some regexp from the injected conf.yaml
-    # to be used later in the verification against the generated terraform.tfvars
-    regexp_set = []
-    conf_dict = yaml.safe_load(conf)
-    for key, value in conf_dict['terraform']['variables'].items():
-        # just focus on the strings variables
-        if isinstance(value, str):
-            regexp_set.append(rf'{key}\s?=\s?"{value}"')
-
-    args, tfvar_path, *_ = args_helper(provider, conf, None)
-    args.append('configure')
-    tfvar_file = os.path.join(tfvar_path, 'terraform.tfvars')
-
-    assert main(args) == 0
-
-    assert os.path.isfile(tfvar_file)
-    with open(tfvar_file, 'r', encoding="utf-8") as file:
-        tfvars_lines = file.readlines()
-        for var_re in regexp_set:
-            one_match = False
-            for line in tfvars_lines:
-                if not one_match:
-                    match = re.search(var_re, line)
-                    if match:
-                        log.debug("Line [%s] match with [%s]", line, var_re)
-                        one_match = True
-            assert one_match, 'Variable:' + var_re + ' missing in the generated terraform.tfvars'
-
-
-def test_configure_create_tfvars_file(configure_helper, config_yaml_sample):
-    """
-    Test that 'configure' write a terraform.tfvars file in
-    <BASE_DIR>/terraform/<PROVIDER>
-    """
-    provider = 'pinocchio'
-    conf = config_yaml_sample(provider)
-    args, tfvar_file, *_ = configure_helper(provider, conf, [])
-
-    assert main(args) == 0
-
-    assert os.path.isfile(tfvar_file)
-
-
-def test_configure_tfvars_novariables_notemplate(configure_helper):
-    """
-    If no terraform.tfvars.template is present and
-    no terraform::variables is present in the config.yaml
-    it has to fails.
-    """
-    provider = 'pinocchio'
-
-    conf = f"""---
-apiver: 2
-provider: {provider}
-ansible:
-    hana_urls: something"""
-    args, *_ = configure_helper(provider, conf, None)
-
-    assert main(args) == 1
-
-
-def test_configure_tfvars_novariables(configure_helper):
-    """
-    Test that 'configure' generated terraform.tfvars file
-    content is like terraform.tfvars.template
-    if no variables are provided in the config.yaml
-    """
-    provider = 'pinocchio'
-    tfvar_template = [
-        "something = static\n",
-        "hananame = hahaha\n",
-        "ip_range = 10.0.4.0/24\n"]
-
-    conf = f"""---
-apiver: 2
-provider: {provider}
-terraform:
-ansible:
-    hana_urls: something"""
-    args, tfvar_path, *_ = configure_helper(provider, conf, tfvar_template)
-
-    assert main(args) == 0
-
-    with open(tfvar_path, 'r', encoding='utf-8') as file:
-        data = file.readlines()
-        assert tfvar_template == data
-
-    conf = f"""---
-apiver: 2
-provider: {provider}
-ansible:
-    hana_urls: something"""
-    args, tfvar_path, *_ = configure_helper(provider, conf, tfvar_template)
-
-    assert main(args) == 0
-
-    with open(tfvar_path, 'r', encoding='utf-8') as file:
-        data = file.readlines()
-        assert tfvar_template == data
-
-
-def test_configure_tfvars_with_variables(configure_helper):
-    """
-    Test that 'configure' generated terraform.tfvars file
-    content is like terraform.tfvars.template
-    plus all key/value pairs from the variables section in
-    the config.yaml
-    """
-    provider = 'pinocchio'
-    conf = f"""---
-apiver: 2
-provider: {provider}
-terraform:
-  variables:
-    region : eu1
-    deployment_name : "rocket"
-ansible:
-    hana_urls: something"""
-    tfvar_template = [
-        "something = static\n",
-        "hananame = hahaha\n",
-        "ip_range = 10.0.4.0/24"]
-
-    args, tfvar_path, *_ = configure_helper(provider, conf, tfvar_template)
-
-    assert main(args) == 0
-
-    expected_tfvars = tfvar_template[0:2]
-    # EOL is expected to be added in terraform.tfvars
-    # if missing at the end of the template
-    expected_tfvars.append(tfvar_template[2] + '\n')
-    expected_tfvars.append('region = "eu1"\n')
-    expected_tfvars.append('deployment_name = "rocket"\n')
-    with open(tfvar_path, 'r', encoding='utf-8') as file:
-        data = file.readlines()
-        assert expected_tfvars == data
-
-
-def test_configure_tfvars_string_commas(configure_helper):
-    """
-    Terraform.tfvars need commas around all strings variables
-    """
-    provider = 'pinocchio'
-    conf = f"""---
-apiver: 2
-provider: {provider}
-terraform:
-  variables:
-    region : eu1
-    deployment_name : "rocket"
-    os_image: SUSE:sles-sap-15-sp3-byos:gen2:2022.05.05
-    public_key: /root/secret/id_rsa.pub
-ansible:
-    hana_urls: something"""
-    tfvar_template = ["something = static"]
-    args, tfvar_path, *_ = configure_helper(provider, conf, tfvar_template)
-
-    assert main(args) == 0
-
-    # EOL is expected to be added in terraform.tfvars
-    # if missing at the end of the template
-    expected_tfvars = [
-        tfvar_template[0] + '\n',
-        'region = "eu1"\n',
-        'deployment_name = "rocket"\n',
-        'os_image = "SUSE:sles-sap-15-sp3-byos:gen2:2022.05.05"\n',
-        'public_key = "/root/secret/id_rsa.pub"\n'
-    ]
-    with open(tfvar_path, 'r', encoding='utf-8') as file:
-        data = file.readlines()
-        assert expected_tfvars == data
-
-
-def test_configure_tfvars_overwrite_variables(configure_helper):
-    """
-    Test 'configure' generated terraform.tfvars file:
-    if same key pair is both in the terraform.tfvars.template
-    and config.yaml, the YAML content win
-    """
-    provider = 'pinocchio'
-
-    conf = f"""---
-apiver: 2
-provider: {provider}
-terraform:
-  variables:
-    something : yamlrulez
-ansible:
-    hana_urls: something"""
-
-    tfvar_template = [
-        "something = static\n",
-        "somethingelse = keep\n"]
-    args, tfvar_path, *_ = configure_helper(provider, conf, tfvar_template)
-
-    assert main(args) == 0
-
-    expected_tfvars = [
-        "something = yamlrulez\n",
-        "somethingelse = keep\n"]
-    with open(tfvar_path, 'r', encoding='utf-8') as file:
-        data = file.readlines()
-        assert expected_tfvars == data
-
-
-def test_configure_create_ansible_hanamedia(configure_helper, config_yaml_sample):
-    """
-    Test that 'configure' write a hana_media.yaml file in
-    <BASE_DIR>/ansible/playbooks/vars
-    """
-    provider = 'pinocchio'
-    conf = config_yaml_sample(provider)
-    args, _, hana_media, _ = configure_helper(provider, conf, [])
-
-    assert main(args) == 0
-
-    assert os.path.isfile(hana_media)
-
-
-def test_configure_ansible_hanamedia_content(configure_helper):
-    """
-    Test that 'configure' write a hana_media.yaml with
-    expected content
-    """
-    provider = 'pinocchio'
-    conf = f"""---
-apiver: 2
-provider: {provider}
-terraform:
-    variables:
-        az_region: "westeurope"
-ansible:
-  hana_urls:
-    - SAPCAR_URL
-    - SAP_HANA_URL
-    - SAP_CLIENT_SAR_URL"""
-    args, _, hana_media, _ = configure_helper(provider, conf, [])
-    assert main(args) == 0
-
-    with open(hana_media, 'r', encoding='utf-8') as file:
-        data = yaml.load(file, Loader=yaml.FullLoader)
-        assert 'hana_urls' in data
-        assert len(data['hana_urls']) == 3
-        assert 'SAPCAR_URL' in data['hana_urls']
-        assert 'SAP_HANA_URL' in data['hana_urls']
-        assert 'SAP_CLIENT_SAR_URL' in data['hana_urls']
-
-
-def test_configure_create_ansible_hanavars(configure_helper, config_yaml_sample):
-    """
-    Test that 'configure' write a hana_vars.yaml file in
-    <BASE_DIR>/ansible/playbooks/vars
-    """
-    provider = 'pinocchio'
-    conf = config_yaml_sample(provider)
-    args, _, _, hana_vars = configure_helper(provider, conf, [])
-
-    assert main(args) == 0
-
-    assert os.path.isfile(hana_vars)
-
-
-def test_configure_not_create_ansible_hanavars_apiver1(configure_helper, config_yaml_sample):
-    """
-    Test that 'configure' does not write a hana_vars.yaml file in
-    <BASE_DIR>/ansible/playbooks/vars
-    if apiver < 2
-    """
-    this_provider = 'pinocchio'
-    conf = config_yaml_sample(provider=this_provider, apiver=1)
-    args, _, _, hana_vars = configure_helper(this_provider, conf, [])
-
-    assert main(args) == 0
-
-    assert not os.path.isfile(hana_vars)
-
-
-def test_configure_ansible_hanavar_content(configure_helper):
-    """
-    Test that 'configure' write a hana_vars.yaml with
-    expected content
-    """
-    provider = 'pinocchio'
-    conf = f"""---
-apiver: 2
-provider: {provider}
-terraform:
-    variables:
-        az_region: "westeurope"
-ansible:
-  hana_urls:
-    - SAPCAR_URL
-  hana_vars:
-    sap_hana_install_software_directory: /hana/shared/install
-    sap_hana_install_master_password: 'DoNotUseThisPassw0rd'
-    sap_hana_install_sid: 'HDB'
-    sap_hana_install_instance_number: '00'
-    sap_domain: "qe-test.example.com"
-    primary_site: 'goofy'
-    secondary_site: 'miky'
-    zanzara: mosquito
-    Moskito: komar
-    moustique: komarac
-"""
-    args, _, _, hana_vars = configure_helper(provider, conf, [])
-    assert main(args) == 0
-
-    with open(hana_vars, 'r', encoding='utf-8') as file:
-        data = yaml.load(file, Loader=yaml.FullLoader)
-        assert 'zanzara' in data
-        assert 'Moskito' in data
-        assert 'moustique' in data
-        assert data['zanzara'] == 'mosquito'
-        assert data['Moskito'] == 'komar'
-        assert data['moustique'] == 'komarac'
-        assert len(data) == 10
-
-def test_configure_ansible_hanavar_values(configure_helper):
-    """
-    Test about value of mandatory fields
-    sap_hana_install_software_directory: string of path where to find HANA software
-    sap_hana_install_master_password: password for <SID>adm user and databases
-    sap_hana_install_sid: Three character SID of the DB.  See restrictions -> https://launchpad.support.sap.com/#/notes/1979280
-    sap_hana_install_instance_number: two digit instance number
-    sap_domain: FQDN with the actual hostname
-    primary_site: name of the primary 'site' for HANA System Replication
-    secondary_site: name of the secondary 'site' for HANA System Replication
-    """
-    provider = 'pinocchio'
-    conf = """---
-apiver: 2
-provider: {}
-terraform:
-    variables:
-        az_region: "westeurope"
-ansible:
-  hana_urls:
-    - SAPCAR_URL
-  hana_vars:
-    sap_hana_install_software_directory: {}
-    sap_hana_install_master_password: ''
-    sap_hana_install_sid: '{}'
-    sap_hana_install_instance_number: '{}'
-    sap_domain: "qe-test.example.com"
-    primary_site: 'goofy'
-    secondary_site: 'miky'
-"""
-    args, _, _, hana_vars = configure_helper(provider, conf.format(provider, '/aaa/bbb/ccc', 'HDB', '00'), [])
-    assert main(args) == 0
-    args, _, _, hana_vars = configure_helper(provider, conf.format(provider, 'ccc', 'HDB', '00'), [])
-    assert main(args) != 0, "Wrong 'sap_hana_install_software_directory'='ccc' not detected."
-    args, _, _, hana_vars = configure_helper(provider, conf.format(provider, '/aaa/bbb/ccc', 'HD', '00'), [])
-    assert main(args) != 0, "Wrong 'sap_hana_install_sid'='HD' not detected."
-    args, _, _, hana_vars = configure_helper(provider, conf.format(provider, '/aaa/bbb/ccc', 'HDB', '0'), [])
-    assert main(args) != 0, "Wrong 'sap_hana_install_instance_number'='0' not detected."
-    args, _, _, hana_vars = configure_helper(provider, conf.format(provider, '/aaa/bbb/ccc', 'HDB', 'AA'), [])
-    assert main(args) != 0, "Wrong 'sap_hana_install_instance_number'='AA' not detected."
-    args, _, _, hana_vars = configure_helper(provider, conf.format(provider, '/aaa/bbb/ccc', 'HDB', '000'), [])
-    assert main(args) != 0, "Wrong 'sap_hana_install_instance_number'='000' not detected."
-
-def test_configure_ansible_hana(configure_helper):
-    """
-    Test that 'configure' fails if manadatory params are missing
-    """
-    provider = 'pinocchio'
-    conf = f"""---
-apiver: 2
-provider: {provider}
-terraform:
-    variables:
-        az_region: "westeurope"
-ansible:
-  hana_urls:
-    - SAPCAR_URL
-  hana_vars:
-    zanzara: mosquito
-"""
-    args, _, _, hana_vars = configure_helper(provider, conf, [])
-    assert main(args) != 0
 
 
 def test_configure_dryrun(config_yaml_sample, configure_helper):
